third_party/WebKit/LayoutTests/netinfo/estimate-multiple-frames.html

<!DOCTYPE html>
<head>
<script src="../resources/js-test.js"></script>
<script src="resources/netinfo_common.js"></script>
</head>
<body>
<script>

description('Tests using NetInfo from multiple frames.');

shouldBe('typeof window.internals.observeGC', '"function"',
'this test requires window.internals');

var childFrame = document.createElement('iframe');
document.body.appendChild(childFrame);

var childConnection = childFrame.contentWindow.navigator.connection;

if (connection.effectiveType != childConnection.effectiveType)
    testFailed("Effective type not the same between main frame and child.");

// Up to 10% noise  may be added to rtt and downlink. rtt or downlink from
// two different hosts may have a difference of up to 20%. Use 21% as the
// buffer below to include any mismatches due to floating point
// calculations.
if (Math.abs(connection.rtt - childConnection.rtt) > 0.21 * childConnection.rtt)
    testFailed("RTT not the same between main frame and child.");
if (Math.abs(connection.downlink - childConnection.downlink) >
       0.21 * childConnection.downlink) {
    testFailed("Downlink not the same between main frame and child.");
}

var hasMainFrameEventFired = false;
var hasChildFrameEventFired = false;

function mainFrameListener() {
    hasMainFrameEventFired = true;
    if (connection.effectiveType != newEffectiveType)
        testFailed("Event fired but effectiveType not yet changed.");
    if (Math.abs(connection.rtt - newRtt) > 0.21 * newRtt)
        testFailed("Event fired but rtt not yet changed.");
    if (Math.abs(connection.downlink - newDownlink) > 0.21 * newDownlink)
        testFailed("Event fired but downlink not yet changed.");
    if (!hasChildFrameEventFired && childConnection.rtt != initialRtt)
        testFailed("Child frame rtt changed before firing its event.");
    maybeFinishTest();
}

function childFrameListener() {
    hasChildFrameEventFired = true;
    if (childConnection.effectiveType != newEffectiveType)
        testFailed("Child frame fired event but effectiveType not yet changed.");
    if (Math.abs(childConnection.rtt - newRtt) > 0.2 * newRtt)
        testFailed("Child frame fired event but rtt not yet changed.");
    if (Math.abs(childConnection.downlink - newDownlink) > 0.2 * newDownlink)
        testFailed("Child frame fired event but downlink not yet changed.");
    if (!hasMainFrameEventFired && connection.rtt != initialRtt)
        testFailed("Main frame rtt changed before firing its event.");
    maybeFinishTest();
}

function maybeFinishTest() {
    if (hasMainFrameEventFired && hasChildFrameEventFired) {
        finishJSTest();
    }
}

connection.addEventListener('change', mainFrameListener);
childConnection.addEventListener('change', childFrameListener);

internals.setNetworkQualityInfoOverride(newEffectiveType, newRtt, newDownlink);

</script>
</body>
</html>
